--- js/src/Dict.js
@@ -1,2 +1,0 @@
-export default class Dict {
-}
