--- js/src/Root.js
@@ -1,11 +1,0 @@
-'use strict';
-import Dict from './Dict.js';
-
-export default class Root {
-    constructor() {
-        this.claims = new Dict();
-        this.scores = new Dict();
-        this.settings = {};
-    }
-}
-//# sourceMappingURL=Root.js.map
