--- js/src/score.js
@@ -1,10 +1,0 @@
-'use strict';
-
-export default class Score {
-    /** */
-    constructor(claim) {
-        if (claim)
-            this.claimId = claim.claimId;
-    }
-}
-//# sourceMappingURL=score.js.map
