+++ package.json
@@ -7,6 +7,7 @@
     "test": "tests"
   },
   "dependencies": {
+    "extract-text-webpack-plugin": "^2.1.2",
     "firebase": "^4.1.2",
     "firebaseui": "^2.1.1",
     "hyperhtml": "^0.11.11",
