+++ public/index.html
@@ -15,9 +15,11 @@
   <script src="js/src/SettleIt.js"></script>
   <script src="https://www.gstatic.com/firebasejs/4.1.1/firebase.js"></script>
   <script src="https://cdn.firebase.com/libs/firebaseui/2.0.0/firebaseui.js"></script>
+  <link rel="stylesheet" type="text/css" href="style.css" media="screen"/>
   <link type="text/css" rel="stylesheet" href="https://cdn.firebase.com/libs/firebaseui/2.0.0/firebaseui.css" /-->
 
   <script src="bundle.js"></script>
+
 
 </head>
