+++ src/app/score.ts
